src/requirements.txt
fastapi
uvicorn
websockets
protobuf
python-multipart
pydantic
einops
openai>=1.0.0
anthropic>=0.3.0
replicate>=0.22.0
accelerate>=1.4.0
transformers>=4.52.4
diffusers>=0.32.2
peft>=0.10.0
huggingface_hub
python-dotenv>=0.19.0
requests>=2.26.0
pillow>=9.0.0
numpy>=1.21.0
scipy>=1.7.0
networkx>=2.6.0
matplotlib>=3.4.0
pyzmq>=22.0.0
colorama>=0.4.4
perlin-noise>=1.8.0
sentencepiece>=0.1.96
sentence-transformers>=2.2.0
python-jose>=3.3.0
aiofiles>=0.8.0
cohere>=3.0.0
google-genai
noise
faiss-gpu-cu12
fastapi[standard]
coqui-tts
hf_xet
vllm>=0.9.1
